Created MockFareCalculator, changed double fare back to int, included fake data for testing, created string resources

+++ app/src/main/java/comcmput301f16t01/github/carrier/FareCalculator.java
@@ -16,7 +16,7 @@ public class FareCalculator {
         // TODO calculate a fare. (this is not for project part 2)
     }
 
-    public double getEstimate() {
+    public int getEstimate() {
         return 0;
     }
 }
